refactor(react-hooks): refactor components/StyleGuide
